+++ src/components/DebugDisplayObj.tsx
